Don't wait for a 401 challenge to pass the auth token to Coconut APIs
In httplib2, `add_credentials()` provides `Basic` authentication only when
the server requests it by returning a 401 challenge.
This makes sense, as an instance of this class can be used to make multiple
requests, not all of which may require auth.

I don't think there's an API in httplib2 to always provide auth,
but we can just do it ourselves by manually specifying an `Authentication` header.

A couple other cleanup changes in here too:
- Make `api_get` and `submit` add the auth token in the same way:
  both use a shared function to check for the auth_token in a parameter first,
  then fall back to the environment variable if it's not specified.
  Before, `api_get` did this, but `submit` did the reverse.
  This may result in slightly different behavior if some client passed one
  auth token as a parameter and a different one in a header
- Raise an assertion if an auth token is not provided,
  since I assume all Coconut APIs require auth.